packages/backend/requirements.txt
# Existing backend requirements
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Database and async
asyncpg>=0.29.0
aiopg>=1.4.0
httpx>=0.25.0

# Environment and config
python-dotenv>=1.0.0
pydantic-settings>=2.1.0

# Audio processing
librosa>=0.10.1
soundfile>=0.12.1

# Phidata AI Agent Framework
phidata>=2.4.0

# LLM Provider SDKs (Updated versions for Phidata compatibility)
openai>=1.52.0
anthropic>=0.30.0
groq>=0.9.0

# Data processing for agents
pandas>=2.1.0
numpy>=1.25.0

# JSON and async utilities
orjson>=3.9.0
aiofiles>=23.2.0

# Monitoring and logging
structlog>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.9.0
isort>=5.12.0httpx>=0.25.0
websockets>=11.0
httpx>=0.25.0
websockets>=11.0
httpx>=0.25.0
websockets>=11.0
httpx>=0.25.0
websockets>=11.0
